focoProject/requirements.txt
django==5.1.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
whitenoise==6.7.0
python-docx==1.1.2